+++ karma.conf.js
@@ -1,11 +1,15 @@
+const puppeteer = require("puppeteer");
+
 process.env.TEST = true;
 process.env.NODE_ENV = "test";
+process.env.CHROME_BIN = puppeteer.executablePath();
 
 const webpackConfig = require("./webpack.config.js")({ production: false, karma: true });
 
 delete webpackConfig.entry;
 
 module.exports = config => {
+  "use strict";
 
   var configuration = {
     client: {
@@ -16,20 +20,19 @@ module.exports = config => {
     basePath: "",
     frameworks: [
       "mocha",
-      "chai",
-      "sinon",
+      "sinon-chai",
       "es6-shim"
     ],
     files: [
       { pattern: "node_modules/reflect-metadata/Reflect.js", include: true },
       { pattern: "node_modules/bluebird/js/browser/bluebird.js", include: true },
       { pattern: "node_modules/phaser/dist/phaser.js", include: true },
-      { pattern: "./test/**/**/**.test.ts", include: true },
+      { pattern: "./test/index.ts", include: true },
       { pattern: '**/*.map', served: true, included: false, watched: true }
     ],
     preprocessors: {
-      "./**/**/**/**.ts": ["sourcemap"],
-      "./test/**/**/**.test.ts": ["webpack"]
+      "./test/index.ts": ["webpack"],
+      "./**/**/**/**.ts": ["sourcemap"]
     },
     webpack: webpackConfig,
     webpackMiddleware: {
@@ -42,11 +45,13 @@ module.exports = config => {
       "karma-remap-istanbul",
       "karma-mocha-reporter",
       "karma-mocha",
-      "karma-chai",
-      "karma-sinon",
+      "karma-sinon-chai",
       "karma-es6-shim",
       "karma-coverage-istanbul-reporter"
     ],
+    mime: {
+      "text/x-typescript": ["ts","tsx"]
+    },
     reporters: (
       config.singleRun ?
         ["dots", "mocha", "coverage-istanbul"] :
@@ -72,11 +77,11 @@ module.exports = config => {
   };
 
   if (process.env.TRAVIS) {
-    configuration.browsers.push("PhantomJS");
-    configuration.plugins.push("karma-phantomjs-launcher");
+    configuration.browsers.push("ChromeHeadless");
+    configuration.plugins.push("karma-chrome-launcher");
   } else {
-    configuration.browsers.push("PhantomJS");
-    configuration.plugins.push("karma-phantomjs-launcher");
+    configuration.browsers.push("ChromeHeadless");
+    configuration.plugins.push("karma-chrome-launcher");
   }
 
   config.set(configuration);